refactor(nodes): split PrimaryNode into focused services (Issue #435)
## Problem

primary-node.ts had 1124 lines with 10+ responsibilities, violating the Single Responsibility Principle:
- WebSocket server management
- HTTP server management
- Execution node management
- Channel management
- Local execution
- Scheduler management
- Task flow orchestration
- File storage service
- Feedback handling
- Command handling

## Solution

Split into focused services following the pattern established in pilot.ts (which uses SessionManager, ConversationOrchestrator, RestartManager):

| Service | Responsibility |
|---------|----------------|
| ExecNodeRegistry | Execution node registration and routing |
| SchedulerService | Scheduler and file watcher lifecycle |
| FeedbackRouter | Feedback routing to channels |
| WebSocketServerService | WebSocket and HTTP server management |

## Changes

| File | Description |
|------|-------------|
| src/nodes/exec-node-registry.ts | New service for execution node management |
| src/nodes/scheduler-service.ts | New service for scheduler lifecycle |
| src/nodes/feedback-router.ts | New service for feedback routing |
| src/nodes/websocket-server-service.ts | New service for server management |
| src/nodes/primary-node.ts | Refactored to use new services (now ~600 lines) |

## Benefits

- Clear separation of concerns
- Easier to test individual services
- Lower maintenance cost
- Better scalability
- Follows established patterns from pilot.ts

## Test Results

| Metric | Value |
|--------|-------|
| Tests | 1210 passed, 8 skipped |
| Type Check | Pass |
| Lint | 0 errors (62 warnings) |

Closes #435
