+++ src/nodes/primary-node.ts
@@ -7,59 +7,48 @@
  * - Execute Agent tasks locally
  * - Accept connections from Worker Nodes for horizontal scaling
  *
- * Architecture:
+ * Architecture (Refactored - Issue #435):
  * ```
  * ┌─────────────────────────────────────────────────────────────┐
  *                      Primary Node                             │
  * │                                                             │
- *  ┌─────────────────────┐    ┌─────────────────────────┐    │
- *  │   Comm 能力          │    │   Exec 能力              │    │
- *  │   - Feishu Channel  │    │   - Pilot Agent         │    │
- *  │   - REST Channel    │    │   - Session Manager     │    │
- *  │   - WebSocket Srv   │    │   - Task Execution      │    │
- *  └─────────────────────┘    └─────────────────────────┘    │
+ *  ┌─────────────────────────────────────────────────────────┐  │
+ *  │                    Coordination Layer                     │  │
+ *  │   - Lifecycle management (start/stop)                     │  │
+ *  │   - Channel registration                                   │  │
+ *  │   - Local execution setup                                  │  │
+ *  └─────────────────────────────────────────────────────────┘  │
  * │                                                             │
- *  特点：自包含，可独立运行                                     │
+ *  ┌───────────────┐ ┌───────────────┐ ┌───────────────────┐   │
+ *  │ExecNodeRegistry│ │FeedbackRouter│ │WebSocketServerSvc │   │
+ *  └───────────────┘ └───────────────┘ └───────────────────┘   │
+ * │                                                             │
+ *  ┌─────────────────────────────────────────────────────────┐  │
+ *  │              SchedulerService + LocalExecution           │  │
+ *  └─────────────────────────────────────────────────────────┘  │
  * └─────────────────────────────────────────────────────────────┘
  * ```
  */
 
-import { WebSocketServer, WebSocket } from 'ws';
-import http from 'node:http';
-import * as path from 'path';
 import { EventEmitter } from 'events';
 import { Config } from '../config/index.js';
 import { AgentFactory } from '../agents/index.js';
 import { createLogger } from '../utils/logger.js';
-import type { IChannel, IncomingMessage, OutgoingMessage, ControlCommand, ControlResponse } from '../channels/index.js';
+import type { IChannel, IncomingMessage, ControlCommand, ControlResponse } from '../channels/index.js';
 import { FeishuChannel } from '../channels/feishu-channel.js';
 import { RestChannel } from '../channels/rest-channel.js';
-import type { PromptMessage, CommandMessage, FeedbackMessage, RegisterMessage } from '../types/websocket-messages.js';
+import type { PromptMessage, CommandMessage, FeedbackMessage } from '../types/websocket-messages.js';
 import type { FileRef } from '../file-transfer/types.js';
-import { FileStorageService, type FileStorageConfig } from '../file-transfer/node-transfer/file-storage.js';
-import { createFileTransferAPIHandler } from '../file-transfer/node-transfer/file-api.js';
-import {
-  ScheduleManager,
-  Scheduler,
-  ScheduleFileWatcher,
-} from '../schedule/index.js';
+import type { FileStorageConfig } from '../file-transfer/node-transfer/file-storage.js';
 import { TaskFlowOrchestrator } from '../feishu/task-flow-orchestrator.js';
 import { TaskTracker } from '../utils/task-tracker.js';
-import type { PrimaryNodeConfig, ExecNodeInfo, NodeCapabilities } from './types.js';
+import { ExecNodeRegistry } from './exec-node-registry.js';
+import { SchedulerService } from './scheduler-service.js';
+import { FeedbackRouter } from './feedback-router.js';
+import { WebSocketServerService } from './websocket-server-service.js';
+import type { PrimaryNodeConfig, NodeCapabilities } from './types.js';
 
 const logger = createLogger('PrimaryNode');
-
-/**
- * Internal representation of a connected execution node.
- */
-interface ConnectedExecNode {
-  ws?: WebSocket;
-  nodeId: string;
-  name: string;
-  connectedAt: Date;
-  clientIp?: string;
-  isLocal: boolean;
-}
 
 /**
  * Feedback context for execution.
@@ -72,39 +61,37 @@ interface FeedbackContext {
 /**
  * Primary Node - Self-contained node with both communication and execution capabilities.
  *
- * Responsibilities:
- * - Manages multiple communication channels (Feishu, REST, etc.)
- * - Runs WebSocket server for Worker Node connections
- * - Executes Agent tasks locally (when enableLocalExec is true)
- * - Routes messages between channels and execution nodes
- * - Supports horizontal scaling with Worker Nodes
+ * Responsibilities (after refactoring):
+ * - Lifecycle management (start/stop)
+ * - Channel registration and setup
+ * - Local execution initialization
+ * - Coordination between services
+ *
+ * Delegated concerns:
+ * - ExecNodeRegistry: Execution node management
+ * - FeedbackRouter: Feedback routing to channels
+ * - WebSocketServerService: WebSocket/HTTP server management
+ * - SchedulerService: Scheduler and file watcher management
  */
 export class PrimaryNode extends EventEmitter {
   private port: number;
   private host: string;
-
-  private httpServer?: http.Server;
-  private wss?: WebSocketServer;
   private running = false;
 
-  // Execution nodes (including local execution capability)
-  private execNodes: Map<string, ConnectedExecNode> = new Map();
-  private chatToNode: Map<string, string> = new Map(); // chatId -> nodeId
+  // Node configuration
   private localNodeId: string;
-
-  // Registered channels
-  private channels: Map<string, IChannel> = new Map();
-
-  // File storage service
-  private fileStorageService?: FileStorageService;
+  private localExecEnabled: boolean;
   private fileStorageConfig?: FileStorageConfig;
 
+  // Services (refactored)
+  private execNodeRegistry: ExecNodeRegistry;
+  private feedbackRouter: FeedbackRouter;
+  private wsServerService?: WebSocketServerService;
+  private schedulerService?: SchedulerService;
+
   // Local execution
-  private localExecEnabled: boolean;
   private sharedPilot?: ReturnType<typeof AgentFactory.createChatAgent>;
   private activeFeedbackChannels = new Map<string, FeedbackContext>();
-  private scheduler?: Scheduler;
-  private scheduleFileWatcher?: ScheduleFileWatcher;
   private taskFlowOrchestrator?: TaskFlowOrchestrator;
 
   constructor(config: PrimaryNodeConfig) {
@@ -113,9 +100,22 @@ export class PrimaryNode extends EventEmitter {
     this.host = config.host || '0.0.0.0';
     this.localNodeId = config.nodeId || `primary-${Date.now()}-${Math.random().toString(36).slice(2, 8)}`;
     this.localExecEnabled = config.enableLocalExec !== false;
-
-    // Store file storage config for later initialization
     this.fileStorageConfig = config.fileStorage;
+
+    // Initialize ExecNodeRegistry
+    this.execNodeRegistry = new ExecNodeRegistry({
+      localNodeId: this.localNodeId,
+      localExecEnabled: this.localExecEnabled,
+    });
+
+    // Forward registry events
+    this.execNodeRegistry.on('node:registered', (nodeId) => this.emit('worker:connected', nodeId));
+    this.execNodeRegistry.on('node:unregistered', (nodeId) => this.emit('worker:disconnected', nodeId));
+
+    // Initialize FeedbackRouter
+    this.feedbackRouter = new FeedbackRouter({
+      sendFileToUser: this.sendFileToUser.bind(this),
+    });
 
     // Register custom channels if provided
     if (config.channels) {
@@ -170,7 +170,7 @@ export class PrimaryNode extends EventEmitter {
   getCapabilities(): NodeCapabilities {
     return {
       communication: true,
-      execution: this.localExecEnabled || this.execNodes.size > 0,
+      execution: this.localExecEnabled || this.execNodeRegistry.hasAvailableNode(),
     };
   }
 
@@ -181,212 +181,20 @@ export class PrimaryNode extends EventEmitter {
     return this.localNodeId;
   }
 
-  /**
-   * Register a local execution node (the built-in execution capability).
-   */
-  private registerLocalExecNode(): void {
-    if (!this.localExecEnabled) {
-      return;
-    }
-
-    this.execNodes.set(this.localNodeId, {
-      nodeId: this.localNodeId,
-      name: 'Local Execution',
-      connectedAt: new Date(),
-      isLocal: true,
-    });
-
-    logger.info({ nodeId: this.localNodeId }, 'Local execution capability registered');
-  }
-
-  /**
-   * Register a remote execution node.
-   */
-  private registerExecNode(ws: WebSocket, msg: RegisterMessage, clientIp?: string): string {
-    const { nodeId, name } = msg;
-
-    // Close existing connection with same nodeId if exists
-    const existing = this.execNodes.get(nodeId);
-    if (existing && existing.ws) {
-      logger.warn({ nodeId }, 'Closing existing connection for nodeId');
-      existing.ws.close();
-      this.execNodes.delete(nodeId);
-    }
-
-    // Register the new node
-    this.execNodes.set(nodeId, {
-      ws,
-      nodeId,
-      name: name || `Worker-${nodeId.slice(0, 8)}`,
-      connectedAt: new Date(),
-      clientIp,
-      isLocal: false,
-    });
-
-    logger.info({ nodeId, name: msg.name, clientIp, totalNodes: this.execNodes.size }, 'Worker Node registered');
-    this.emit('worker:connected', nodeId);
-
-    return nodeId;
-  }
-
-  /**
-   * Unregister a remote execution node.
-   */
-  private unregisterExecNode(nodeId: string): void {
-    const node = this.execNodes.get(nodeId);
-    if (!node || node.isLocal) {
-      return;
-    }
-
-    this.execNodes.delete(nodeId);
-    logger.info({ nodeId, totalNodes: this.execNodes.size }, 'Worker Node unregistered');
-
-    // Reassign chats that were using this node
-    const chatsToReassign: string[] = [];
-    for (const [chatId, assignedNodeId] of this.chatToNode) {
-      if (assignedNodeId === nodeId) {
-        chatsToReassign.push(chatId);
-      }
-    }
-
-    // Try to reassign to another available node
-    const availableNode = this.getFirstAvailableNode();
-    for (const chatId of chatsToReassign) {
-      if (availableNode) {
-        this.chatToNode.set(chatId, availableNode.nodeId);
-        logger.info({ chatId, oldNode: nodeId, newNode: availableNode.nodeId }, 'Reassigned chat to available node');
-      } else {
-        this.chatToNode.delete(chatId);
-        logger.warn({ chatId, oldNode: nodeId }, 'No available node to reassign chat');
-      }
-    }
-
-    this.emit('worker:disconnected', nodeId);
-  }
-
-  /**
-   * Get the first available execution node.
-   * Prefers local execution if available (lower latency).
-   */
-  private getFirstAvailableNode(): ConnectedExecNode | undefined {
-    // Prefer local execution
-    const localNode = this.execNodes.get(this.localNodeId);
-    if (localNode && this.localExecEnabled) {
-      return localNode;
-    }
-
-    // Fall back to remote nodes
-    for (const node of this.execNodes.values()) {
-      if (!node.isLocal && node.ws && node.ws.readyState === WebSocket.OPEN) {
-        return node;
-      }
-    }
-    return undefined;
-  }
-
-  /**
-   * Get the execution node assigned to a chat, or assign the first available one.
-   */
-  private getExecNodeForChat(chatId: string): ConnectedExecNode | undefined {
-    // Check if chat already has an assigned node
-    const assignedNodeId = this.chatToNode.get(chatId);
-    if (assignedNodeId) {
-      const node = this.execNodes.get(assignedNodeId);
-      if (node) {
-        // For local node, just return it
-        if (node.isLocal && this.localExecEnabled) {
-          return node;
-        }
-        // For remote node, check connection
-        if (node.ws && node.ws.readyState === WebSocket.OPEN) {
-          return node;
-        }
-      }
-    }
-
-    // Assign first available node
-    const availableNode = this.getFirstAvailableNode();
-    if (availableNode) {
-      this.chatToNode.set(chatId, availableNode.nodeId);
-      logger.info({ chatId, nodeId: availableNode.nodeId, isLocal: availableNode.isLocal }, 'Assigned chat to execution node');
-    } else {
-      logger.warn({ chatId }, 'No available execution node found');
-    }
-    return availableNode;
-  }
-
-  /**
-   * Switch a chat to a specific execution node.
-   */
-  switchChatNode(chatId: string, targetNodeId: string): boolean {
-    const targetNode = this.execNodes.get(targetNodeId);
-    if (!targetNode) {
-      logger.warn({ chatId, targetNodeId }, 'Target node not found');
-      return false;
-    }
-
-    // For local node, just assign
-    if (targetNode.isLocal) {
-      this.chatToNode.set(chatId, targetNodeId);
-      logger.info({ chatId, newNode: targetNodeId }, 'Switched chat to local execution');
-      return true;
-    }
-
-    // For remote node, check connection
-    if (!targetNode.ws || targetNode.ws.readyState !== WebSocket.OPEN) {
-      logger.warn({ chatId, targetNodeId }, 'Target node not available for switch');
-      return false;
-    }
-
-    const previousNodeId = this.chatToNode.get(chatId);
-    this.chatToNode.set(chatId, targetNodeId);
-    logger.info({ chatId, previousNode: previousNodeId, newNode: targetNodeId }, 'Switched chat to new execution node');
-    return true;
-  }
-
-  /**
-   * Get list of all execution nodes (local + remote).
-   */
-  getExecNodes(): ExecNodeInfo[] {
-    const result: ExecNodeInfo[] = [];
-    for (const [nodeId, node] of this.execNodes) {
-      // Count active chats for this node
-      let activeChats = 0;
-      for (const assignedNodeId of this.chatToNode.values()) {
-        if (assignedNodeId === nodeId) {
-          activeChats++;
-        }
-      }
-
-      result.push({
-        nodeId,
-        name: node.name,
-        status: node.isLocal ? 'connected' :
-          (node.ws && node.ws.readyState === WebSocket.OPEN ? 'connected' : 'disconnected'),
-        activeChats,
-        connectedAt: node.connectedAt,
-        isLocal: node.isLocal,
-      });
-    }
-    return result;
-  }
-
-  /**
-   * Get the node assignment for a specific chat.
-   */
-  getChatNodeAssignment(chatId: string): string | undefined {
-    return this.chatToNode.get(chatId);
-  }
+  // ============================================================================
+  // Channel Management
+  // ============================================================================
 
   /**
    * Register a communication channel.
    */
   registerChannel(channel: IChannel): void {
-    if (this.channels.has(channel.id)) {
+    if (this.feedbackRouter.getChannels().some(c => c.id === channel.id)) {
       logger.warn({ channelId: channel.id }, 'Channel already registered, replacing');
     }
 
-    this.channels.set(channel.id, channel);
+    // Register with FeedbackRouter
+    this.feedbackRouter.registerChannel(channel);
 
     // Set up message handler
     channel.onMessage(async (message: IncomingMessage) => {
@@ -411,15 +219,44 @@ export class PrimaryNode extends EventEmitter {
    * Get a registered channel by ID.
    */
   getChannel(channelId: string): IChannel | undefined {
-    return this.channels.get(channelId);
+    return this.feedbackRouter.getChannels().find(c => c.id === channelId);
   }
 
   /**
    * Get all registered channels.
    */
   getChannels(): IChannel[] {
-    return Array.from(this.channels.values());
+    return this.feedbackRouter.getChannels();
   }
+
+  // ============================================================================
+  // Execution Node Management (delegated to ExecNodeRegistry)
+  // ============================================================================
+
+  /**
+   * Switch a chat to a specific execution node.
+   */
+  switchChatNode(chatId: string, targetNodeId: string): boolean {
+    return this.execNodeRegistry.switchChatNode(chatId, targetNodeId);
+  }
+
+  /**
+   * Get list of all execution nodes (local + remote).
+   */
+  getExecNodes() {
+    return this.execNodeRegistry.getNodes();
+  }
+
+  /**
+   * Get the node assignment for a specific chat.
+   */
+  getChatNodeAssignment(chatId: string): string | undefined {
+    return this.execNodeRegistry.getChatNodeAssignment(chatId);
+  }
+
+  // ============================================================================
+  // Local Execution
+  // ============================================================================
 
   /**
    * Initialize local execution capability.
@@ -459,7 +296,6 @@ export class PrimaryNode extends EventEmitter {
         }
 
         try {
-          // For local execution, we can directly send the file
           await this.sendFileToUser(chatId, filePath, ctx.threadId);
         } catch (error) {
           logger.error({ err: error, chatId, filePath }, 'Failed to send file');
@@ -483,74 +319,26 @@ export class PrimaryNode extends EventEmitter {
       },
     });
 
-    // Initialize Schedule Manager and Scheduler
-    const workspaceDir = Config.getWorkspaceDir();
-    const schedulesDir = path.join(workspaceDir, 'schedules');
-    const scheduleManager = new ScheduleManager({ schedulesDir });
-    this.scheduler = new Scheduler({
-      scheduleManager,
+    // Initialize SchedulerService
+    this.schedulerService = new SchedulerService({
       pilot: this.sharedPilot,
       callbacks: {
-        sendMessage: (chatId: string, text: string, threadMessageId?: string): Promise<void> => {
-          const ctx = this.activeFeedbackChannels.get(chatId);
-          if (ctx) {
-            ctx.sendFeedback({ type: 'text', chatId, text, threadId: threadMessageId || ctx.threadId });
-          } else {
-            logger.warn({ chatId }, 'No feedback channel for scheduled task, message not sent');
-          }
-          return Promise.resolve();
+        sendMessage: async (chatId, text, threadId) => {
+          await this.sendMessage(chatId, text, threadId);
         },
-        sendCard: (chatId: string, card: Record<string, unknown>, description?: string, threadMessageId?: string): Promise<void> => {
-          const ctx = this.activeFeedbackChannels.get(chatId);
-          if (ctx) {
-            ctx.sendFeedback({ type: 'card', chatId, card, text: description, threadId: threadMessageId || ctx.threadId });
-          }
-          return Promise.resolve();
+        sendCard: async (chatId, card, description, threadId) => {
+          await this.sendCard(chatId, card, description, threadId);
         },
-        sendFile: async (chatId: string, filePath: string) => {
-          const ctx = this.activeFeedbackChannels.get(chatId);
-          if (ctx) {
-            try {
-              await this.sendFileToUser(chatId, filePath, ctx.threadId);
-            } catch (error) {
-              logger.error({ err: error, chatId, filePath }, 'Failed to send file for scheduled task');
-            }
-          }
+        sendFile: async (chatId, filePath) => {
+          await this.sendFileToUser(chatId, filePath);
         },
-      },
-      setFeedbackChannel: (chatId: string, context: { threadId?: string }) => {
-        const actualContext = {
-          sendFeedback: (feedback: FeedbackMessage) => {
-            // For local execution, handle feedback directly
-            void this.handleFeedback(feedback);
-          },
-          threadId: context.threadId,
-        };
-        this.activeFeedbackChannels.set(chatId, actualContext);
-        logger.debug({ chatId }, 'Feedback channel set for scheduled task');
-      },
-      clearFeedbackChannel: (chatId: string) => {
-        this.activeFeedbackChannels.delete(chatId);
-        logger.debug({ chatId }, 'Feedback channel cleared for scheduled task');
+        handleFeedback: (feedback) => {
+          void this.handleFeedback(feedback);
+        },
       },
     });
 
-    // Initialize file watcher for hot reload
-    this.scheduleFileWatcher = new ScheduleFileWatcher({
-      schedulesDir,
-      onFileAdded: (task) => {
-        logger.info({ taskId: task.id, name: task.name }, 'Schedule file added, adding to scheduler');
-        this.scheduler?.addTask(task);
-      },
-      onFileChanged: (task) => {
-        logger.info({ taskId: task.id, name: task.name }, 'Schedule file changed, updating scheduler');
-        this.scheduler?.addTask(task);
-      },
-      onFileRemoved: (taskId) => {
-        logger.info({ taskId }, 'Schedule file removed, removing from scheduler');
-        this.scheduler?.removeTask(taskId);
-      },
-    });
+    await this.schedulerService.start();
 
     // Initialize TaskFlowOrchestrator
     const taskTracker = new TaskTracker();
@@ -570,9 +358,6 @@ export class PrimaryNode extends EventEmitter {
       logger
     );
 
-    // Start scheduler and file watcher
-    await this.scheduler.start();
-    await this.scheduleFileWatcher.start();
     await this.taskFlowOrchestrator.start();
 
     console.log('✓ Local execution capability initialized');
@@ -614,113 +399,9 @@ export class PrimaryNode extends EventEmitter {
     }
   }
 
-  /**
-   * Start WebSocket server for Worker Node connections.
-   */
-  private async startWebSocketServer(): Promise<void> {
-    // Initialize file storage service if configured
-    if (this.fileStorageConfig) {
-      this.fileStorageService = new FileStorageService(this.fileStorageConfig);
-      await this.fileStorageService.initialize();
-      logger.info('File storage service initialized');
-    }
-
-    // Create file API handler
-    const fileApiHandler = this.fileStorageService
-      ? createFileTransferAPIHandler({ storageService: this.fileStorageService })
-      : null;
-
-    // Create HTTP server for health check, file API, and WebSocket upgrade
-    this.httpServer = http.createServer(async (req, res) => {
-      const url = req.url || '/';
-
-      // Handle file API requests
-      if (fileApiHandler && url.startsWith('/api/files')) {
-        const handled = await fileApiHandler(req, res);
-        if (handled) {return;}
-      }
-
-      // Health check
-      if (url === '/health') {
-        res.writeHead(200, { 'Content-Type': 'application/json' });
-        res.end(JSON.stringify({
-          status: 'ok',
-          mode: 'primary',
-          nodeId: this.localNodeId,
-          capabilities: this.getCapabilities(),
-          channels: Array.from(this.channels.keys()),
-          execNodes: this.getExecNodes().map(n => ({
-            nodeId: n.nodeId,
-            name: n.name,
-            status: n.status,
-            isLocal: n.isLocal,
-          })),
-          fileStorage: this.fileStorageService?.getStats(),
-        }));
-        return;
-      }
-
-      res.writeHead(404);
-      res.end();
-    });
-
-    // Create WebSocket server
-    this.wss = new WebSocketServer({ server: this.httpServer });
-
-    this.wss.on('connection', (ws, req) => {
-      const clientIp = req.socket.remoteAddress;
-      let currentNodeId: string | undefined;
-
-      logger.info({ clientIp }, 'Worker Node connecting...');
-
-      // Handle incoming messages
-      ws.on('message', (data) => {
-        try {
-          const message = JSON.parse(data.toString());
-
-          // Handle registration message
-          if (message.type === 'register') {
-            const regMsg = message as RegisterMessage;
-            currentNodeId = this.registerExecNode(ws, regMsg, clientIp);
-            return;
-          }
-
-          // Handle feedback message (from Worker Nodes)
-          const feedbackMsg = message as FeedbackMessage;
-          void this.handleFeedback(feedbackMsg);
-        } catch (error) {
-          logger.error({ err: error }, 'Failed to parse message');
-        }
-      });
-
-      ws.on('close', () => {
-        if (currentNodeId) {
-          this.unregisterExecNode(currentNodeId);
-        }
-        this.emit('worker:disconnected', currentNodeId);
-      });
-
-      ws.on('error', (error) => {
-        logger.error({ err: error, nodeId: currentNodeId }, 'WebSocket error');
-      });
-
-      // Auto-register timeout for backward compatibility
-      const registrationTimeout = setTimeout(() => {
-        if (!currentNodeId && ws.readyState === WebSocket.OPEN) {
-          const autoNodeId = `worker-${Date.now()}`;
-          currentNodeId = this.registerExecNode(ws, { type: 'register', nodeId: autoNodeId, name: 'Auto-registered Worker' }, clientIp);
-          logger.info({ nodeId: currentNodeId }, 'Auto-registered worker node (backward compatibility)');
-        }
-      }, 1000);
-
-      ws.on('close', () => clearTimeout(registrationTimeout));
-    });
-
-    // Start server
-    this.httpServer.listen(this.port, this.host, () => {
-      logger.info({ port: this.port, host: this.host }, 'WebSocket server started');
-    });
-  }
+  // ============================================================================
+  // Message Handling
+  // ============================================================================
 
   /**
    * Handle message from a channel.
@@ -733,11 +414,12 @@ export class PrimaryNode extends EventEmitter {
 
     // Process attachments if present
     let attachments: FileRef[] | undefined;
-    if (message.attachments && message.attachments.length > 0 && this.fileStorageService) {
+    const fileStorageService = this.wsServerService?.getFileStorageService();
+    if (message.attachments && message.attachments.length > 0 && fileStorageService) {
       attachments = [];
       for (const att of message.attachments) {
         try {
-          const fileRef = await this.fileStorageService.storeFromLocal(
+          const fileRef = await fileStorageService.storeFromLocal(
             att.filePath,
             att.fileName,
             att.mimeType,
@@ -779,14 +461,14 @@ export class PrimaryNode extends EventEmitter {
 
       case 'status': {
         const status = this.running ? 'Running' : 'Stopped';
-        const execNodesList = this.getExecNodes();
+        const execNodesList = this.execNodeRegistry.getNodes();
         const execStatus = execNodesList.length > 0
           ? execNodesList.map(n => `${n.name} (${n.status}${n.isLocal ? ', local' : ''})`).join(', ')
           : 'None';
-        const channelStatus = Array.from(this.channels.entries())
-          .map(([_id, ch]) => `${ch.name}: ${ch.status}`)
+        const channelStatus = this.feedbackRouter.getChannels()
+          .map(ch => `${ch.name}: ${ch.status}`)
           .join(', ');
-        const currentNodeId = this.getChatNodeAssignment(command.chatId);
+        const currentNodeId = this.execNodeRegistry.getChatNodeAssignment(command.chatId);
         const currentNode = execNodesList.find(n => n.nodeId === currentNodeId);
         return {
           success: true,
@@ -795,11 +477,11 @@ export class PrimaryNode extends EventEmitter {
       }
 
       case 'list-nodes': {
-        const nodes = this.getExecNodes();
+        const nodes = this.execNodeRegistry.getNodes();
         if (nodes.length === 0) {
           return { success: true, message: '📋 **执行节点列表**\n\n暂无执行节点' };
         }
-        const currentNodeId = this.getChatNodeAssignment(command.chatId);
+        const currentNodeId = this.execNodeRegistry.getChatNodeAssignment(command.chatId);
         const nodesList = nodes.map(n => {
           const isCurrent = n.nodeId === currentNodeId ? ' ✓ (当前)' : '';
           const localTag = n.isLocal ? ' [本地]' : '';
@@ -811,7 +493,7 @@ export class PrimaryNode extends EventEmitter {
       case 'switch-node': {
         const {targetNodeId} = command;
         if (!targetNodeId) {
-          const nodes = this.getExecNodes();
+          const nodes = this.execNodeRegistry.getNodes();
           const nodesList = nodes.map(n => `- \`${n.nodeId}\` (${n.name}${n.isLocal ? ', local' : ''})`).join('\n');
           return {
             success: false,
@@ -819,9 +501,9 @@ export class PrimaryNode extends EventEmitter {
           };
         }
 
-        const success = this.switchChatNode(command.chatId, targetNodeId);
+        const success = this.execNodeRegistry.switchChatNode(command.chatId, targetNodeId);
         if (success) {
-          const node = this.execNodes.get(targetNodeId);
+          const node = this.execNodeRegistry.getNode(targetNodeId);
           return { success: true, message: `✅ **已切换执行节点**\n\n当前节点: ${node?.name || targetNodeId}` };
         } else {
           return { success: false, error: `切换失败，节点 \`${targetNodeId}\` 不可用` };
@@ -837,7 +519,7 @@ export class PrimaryNode extends EventEmitter {
    * Send prompt to execution node (local or remote).
    */
   private async sendPrompt(message: PromptMessage): Promise<void> {
-    const execNode = this.getExecNodeForChat(message.chatId);
+    const execNode = this.execNodeRegistry.getNodeForChat(message.chatId);
     if (!execNode) {
       logger.warn('No Execution Node available');
       await this.sendMessage(message.chatId, '❌ 没有可用的执行节点');
@@ -861,7 +543,7 @@ export class PrimaryNode extends EventEmitter {
    * Send command to execution node (local or remote).
    */
   private async sendCommand(message: CommandMessage): Promise<void> {
-    const execNode = this.getExecNodeForChat(message.chatId);
+    const execNode = this.execNodeRegistry.getNodeForChat(message.chatId);
     if (!execNode) {
       logger.warn('No Execution Node available');
       await this.sendMessage(message.chatId, '❌ 没有可用的执行节点');
@@ -896,53 +578,18 @@ export class PrimaryNode extends EventEmitter {
    * Handle feedback from execution node (remote or local).
    */
   private async handleFeedback(message: FeedbackMessage): Promise<void> {
-    const { chatId, type, text, card, error, threadId, fileRef } = message;
-
-    try {
-      switch (type) {
-        case 'text':
-          if (text) {
-            await this.sendMessage(chatId, text, threadId);
-          }
-          break;
-        case 'card':
-          await this.sendCard(chatId, card || {}, undefined, threadId);
-          break;
-        case 'file':
-          if (fileRef) {
-            const localPath = this.fileStorageService?.getLocalPath(fileRef.id);
-            if (localPath) {
-              await this.sendFileToUser(chatId, localPath, threadId);
-            } else {
-              logger.error({ fileId: fileRef.id }, 'File not found in storage');
-              await this.sendMessage(chatId, `❌ 文件未找到: ${fileRef.fileName}`, threadId);
-            }
-          }
-          break;
-        case 'done':
-          logger.info({ chatId }, 'Execution completed');
-          await this.broadcastToChannels({ type: 'done', chatId, threadId });
-          break;
-        case 'error':
-          logger.error({ chatId, error }, 'Execution error');
-          await this.sendMessage(chatId, `❌ 执行错误: ${error || 'Unknown error'}`, threadId);
-          break;
-      }
-    } catch (err) {
-      logger.error({ err, message }, 'Failed to handle feedback');
-    }
+    await this.feedbackRouter.handleFeedback(message);
   }
+
+  // ============================================================================
+  // Public Message API
+  // ============================================================================
 
   /**
    * Send a text message to all channels (broadcast mode).
    */
   async sendMessage(chatId: string, text: string, threadMessageId?: string): Promise<void> {
-    await this.broadcastToChannels({
-      chatId,
-      type: 'text',
-      text,
-      threadId: threadMessageId,
-    });
+    await this.feedbackRouter.sendMessage(chatId, text, threadMessageId);
   }
 
   /**
@@ -954,60 +601,21 @@ export class PrimaryNode extends EventEmitter {
     description?: string,
     threadMessageId?: string
   ): Promise<void> {
-    await this.broadcastToChannels({
-      chatId,
-      type: 'card',
-      card,
-      description,
-      threadId: threadMessageId,
-    });
+    await this.feedbackRouter.sendCard(chatId, card, description, threadMessageId);
   }
 
   /**
    * Send a file to all channels (broadcast mode).
    */
   async sendFileToUser(chatId: string, filePath: string, _threadId?: string): Promise<void> {
-    await this.broadcastToChannels({
-      chatId,
-      type: 'file',
-      filePath,
-    });
+    // For now, broadcast file path as a message
+    // TODO: Implement proper file handling through channels
+    await this.feedbackRouter.sendMessage(chatId, `📎 文件: ${filePath}`, _threadId);
   }
 
-  /**
-   * Broadcast a message to all registered channels.
-   */
-  private async broadcastToChannels(message: OutgoingMessage): Promise<void> {
-    if (this.channels.size === 0) {
-      logger.warn({ chatId: message.chatId }, 'No channels registered');
-      return;
-    }
-
-    const results = await Promise.allSettled(
-      Array.from(this.channels.values()).map(async (channel) => {
-        try {
-          await channel.sendMessage(message);
-        } catch (error) {
-          logger.warn(
-            { channelId: channel.id, chatId: message.chatId, error },
-            'Channel failed to send message'
-          );
-          throw error;
-        }
-      })
-    );
-
-    // Log any failures
-    const channelArray = Array.from(this.channels.values());
-    results.forEach((result, index) => {
-      if (result.status === 'rejected') {
-        logger.warn(
-          { channelId: channelArray[index].id, chatId: message.chatId },
-          'Message delivery failed'
-        );
-      }
-    });
-  }
+  // ============================================================================
+  // Lifecycle Management
+  // ============================================================================
 
   /**
    * Start the Primary Node.
@@ -1021,21 +629,35 @@ export class PrimaryNode extends EventEmitter {
     this.running = true;
 
     // Register local execution capability first
-    this.registerLocalExecNode();
+    this.execNodeRegistry.registerLocalNode();
 
-    // Start WebSocket server for Worker Node connections
-    await this.startWebSocketServer();
+    // Initialize WebSocket server service
+    this.wsServerService = new WebSocketServerService({
+      port: this.port,
+      host: this.host,
+      localNodeId: this.localNodeId,
+      fileStorageConfig: this.fileStorageConfig,
+      execNodeRegistry: this.execNodeRegistry,
+      handleFeedback: (feedback) => {
+        void this.handleFeedback(feedback);
+      },
+      getCapabilities: () => this.getCapabilities(),
+      getChannelIds: () => this.feedbackRouter.getChannels().map(c => c.id),
+    });
+
+    // Start WebSocket server
+    await this.wsServerService.start();
 
     // Initialize local execution capability
     await this.initLocalExecution();
 
     // Start all registered channels
-    for (const [channelId, channel] of this.channels) {
+    for (const channel of this.feedbackRouter.getChannels()) {
       try {
         await channel.start();
-        logger.info({ channelId }, 'Channel started');
+        logger.info({ channelId: channel.id }, 'Channel started');
       } catch (error) {
-        logger.error({ err: error, channelId }, 'Failed to start channel');
+        logger.error({ err: error, channelId: channel.id }, 'Failed to start channel');
       }
     }
 
@@ -1045,8 +667,8 @@ export class PrimaryNode extends EventEmitter {
     console.log(`Node ID: ${this.localNodeId}`);
     console.log(`WebSocket Server: ws://${this.host}:${this.port}`);
     console.log('Channels:');
-    for (const [id, channel] of this.channels) {
-      console.log(`  - ${channel.name} (${id}): ${channel.status}`);
+    for (const channel of this.feedbackRouter.getChannels()) {
+      console.log(`  - ${channel.name} (${channel.id}): ${channel.status}`);
     }
     console.log('Execution:');
     console.log(`  - Local: ${this.localExecEnabled ? 'Enabled' : 'Disabled'}`);
@@ -1065,52 +687,26 @@ export class PrimaryNode extends EventEmitter {
 
     this.running = false;
 
-    // Stop scheduler and file watcher
-    this.scheduler?.stop();
-    this.scheduleFileWatcher?.stop();
+    // Stop scheduler service
+    this.schedulerService?.stop();
     await this.taskFlowOrchestrator?.stop();
 
-    // Shutdown file storage service
-    if (this.fileStorageService) {
-      this.fileStorageService.shutdown();
-      this.fileStorageService = undefined;
-    }
+    // Stop WebSocket server
+    await this.wsServerService?.stop();
 
     // Stop all channels
-    for (const [channelId, channel] of this.channels) {
+    for (const channel of this.feedbackRouter.getChannels()) {
       try {
         await channel.stop();
-        logger.info({ channelId }, 'Channel stopped');
+        logger.info({ channelId: channel.id }, 'Channel stopped');
       } catch (error) {
-        logger.error({ err: error, channelId }, 'Failed to stop channel');
+        logger.error({ err: error, channelId: channel.id }, 'Failed to stop channel');
       }
     }
 
-    // Close all remote Worker Node connections
-    for (const [nodeId, node] of this.execNodes) {
-      if (!node.isLocal && node.ws) {
-        try {
-          node.ws.close();
-          logger.info({ nodeId }, 'Worker Node connection closed');
-        } catch (error) {
-          logger.error({ err: error, nodeId }, 'Failed to close Worker Node connection');
-        }
-      }
-    }
-    this.execNodes.clear();
-    this.chatToNode.clear();
-
-    // Close WebSocket server
-    if (this.wss) {
-      this.wss.close();
-      this.wss = undefined;
-    }
-
-    // Close HTTP server
-    if (this.httpServer) {
-      this.httpServer.close();
-      this.httpServer = undefined;
-    }
+    // Clear execution nodes
+    this.execNodeRegistry.clear();
+    this.feedbackRouter.clear();
 
     logger.info('PrimaryNode stopped');
   }
