+++ src/nodes/scheduler-service.ts
@@ -1,0 +1,193 @@
+/**
+ * SchedulerService - Manages scheduler and schedule file watcher.
+ *
+ * Extracts scheduler management concerns from PrimaryNode:
+ * - Scheduler initialization and lifecycle
+ * - Schedule file watching for hot reload
+ * - Callbacks for schedule execution
+ *
+ * Architecture:
+ * ```
+ * PrimaryNode → SchedulerService → { Scheduler, ScheduleFileWatcher }
+ *                      ↓
+ *              ScheduleManager → schedule execution
+ * ```
+ */
+
+import * as path from 'path';
+import { createLogger } from '../utils/logger.js';
+import { Config } from '../config/index.js';
+import {
+  ScheduleManager,
+  Scheduler,
+  ScheduleFileWatcher,
+} from '../schedule/index.js';
+import type { FeedbackMessage } from '../types/websocket-messages.js';
+import type { ChatAgent } from '../agents/types.js';
+
+const logger = createLogger('SchedulerService');
+
+/**
+ * Callbacks for schedule execution.
+ */
+export interface SchedulerCallbacks {
+  /** Send a text message */
+  sendMessage: (chatId: string, text: string, threadMessageId?: string) => Promise<void>;
+  /** Send a card message */
+  sendCard: (chatId: string, card: Record<string, unknown>, description?: string, threadMessageId?: string) => Promise<void>;
+  /** Send a file */
+  sendFile: (chatId: string, filePath: string) => Promise<void>;
+  /** Handle feedback from schedule execution */
+  handleFeedback: (feedback: FeedbackMessage) => void;
+}
+
+/**
+ * Configuration for SchedulerService.
+ */
+export interface SchedulerServiceConfig {
+  /** Callbacks for schedule execution */
+  callbacks: SchedulerCallbacks;
+  /** Pilot agent for schedule execution */
+  pilot: ChatAgent;
+}
+
+/**
+ * Feedback context for execution.
+ */
+interface FeedbackContext {
+  sendFeedback: (feedback: FeedbackMessage) => void;
+  threadId?: string;
+}
+
+/**
+ * SchedulerService - Manages scheduler lifecycle.
+ *
+ * Handles:
+ * - Scheduler initialization
+ * - Schedule file watching
+ * - Feedback channel management
+ */
+export class SchedulerService {
+  private readonly callbacks: SchedulerCallbacks;
+  private readonly pilot: SchedulerServiceConfig['pilot'];
+  private readonly activeFeedbackChannels = new Map<string, FeedbackContext>();
+  private scheduler?: Scheduler;
+  private scheduleFileWatcher?: ScheduleFileWatcher;
+  private schedulesDir: string;
+
+  constructor(config: SchedulerServiceConfig) {
+    this.callbacks = config.callbacks;
+    this.pilot = config.pilot;
+
+    const workspaceDir = Config.getWorkspaceDir();
+    this.schedulesDir = path.join(workspaceDir, 'schedules');
+  }
+
+  /**
+   * Start the scheduler service.
+   */
+  async start(): Promise<void> {
+    const scheduleManager = new ScheduleManager({ schedulesDir: this.schedulesDir });
+
+    this.scheduler = new Scheduler({
+      scheduleManager,
+      pilot: this.pilot,
+      callbacks: {
+        sendMessage: (chatId: string, text: string, threadMessageId?: string): Promise<void> => {
+          const ctx = this.activeFeedbackChannels.get(chatId);
+          if (ctx) {
+            ctx.sendFeedback({ type: 'text', chatId, text, threadId: threadMessageId || ctx.threadId });
+          } else {
+            logger.warn({ chatId }, 'No feedback channel for scheduled task, message not sent');
+          }
+          return Promise.resolve();
+        },
+        sendCard: (chatId: string, card: Record<string, unknown>, description?: string, threadMessageId?: string): Promise<void> => {
+          const ctx = this.activeFeedbackChannels.get(chatId);
+          if (ctx) {
+            ctx.sendFeedback({ type: 'card', chatId, card, text: description, threadId: threadMessageId || ctx.threadId });
+          }
+          return Promise.resolve();
+        },
+        sendFile: async (chatId: string, filePath: string) => {
+          const ctx = this.activeFeedbackChannels.get(chatId);
+          if (ctx) {
+            try {
+              await this.callbacks.sendFile(chatId, filePath);
+            } catch (error) {
+              logger.error({ err: error, chatId, filePath }, 'Failed to send file for scheduled task');
+            }
+          }
+        },
+      },
+      setFeedbackChannel: (chatId: string, context: { threadId?: string }) => {
+        const actualContext = {
+          sendFeedback: (feedback: FeedbackMessage) => {
+            this.callbacks.handleFeedback(feedback);
+          },
+          threadId: context.threadId,
+        };
+        this.activeFeedbackChannels.set(chatId, actualContext);
+        logger.debug({ chatId }, 'Feedback channel set for scheduled task');
+      },
+      clearFeedbackChannel: (chatId: string) => {
+        this.activeFeedbackChannels.delete(chatId);
+        logger.debug({ chatId }, 'Feedback channel cleared for scheduled task');
+      },
+    });
+
+    // Initialize file watcher for hot reload
+    this.scheduleFileWatcher = new ScheduleFileWatcher({
+      schedulesDir: this.schedulesDir,
+      onFileAdded: (task) => {
+        logger.info({ taskId: task.id, name: task.name }, 'Schedule file added, adding to scheduler');
+        this.scheduler?.addTask(task);
+      },
+      onFileChanged: (task) => {
+        logger.info({ taskId: task.id, name: task.name }, 'Schedule file changed, updating scheduler');
+        this.scheduler?.addTask(task);
+      },
+      onFileRemoved: (taskId) => {
+        logger.info({ taskId }, 'Schedule file removed, removing from scheduler');
+        this.scheduler?.removeTask(taskId);
+      },
+    });
+
+    // Start scheduler and file watcher
+    await this.scheduler.start();
+    await this.scheduleFileWatcher.start();
+
+    logger.info('Scheduler service started');
+  }
+
+  /**
+   * Stop the scheduler service.
+   */
+  stop(): void {
+    this.scheduler?.stop();
+    this.scheduleFileWatcher?.stop();
+    this.activeFeedbackChannels.clear();
+    logger.info('Scheduler service stopped');
+  }
+
+  /**
+   * Set feedback channel for a chat.
+   */
+  setFeedbackChannel(chatId: string, sendFeedback: (feedback: FeedbackMessage) => void, threadId?: string): void {
+    this.activeFeedbackChannels.set(chatId, { sendFeedback, threadId });
+  }
+
+  /**
+   * Clear feedback channel for a chat.
+   */
+  clearFeedbackChannel(chatId: string): void {
+    this.activeFeedbackChannels.delete(chatId);
+  }
+
+  /**
+   * Get the scheduler instance.
+   */
+  getScheduler(): Scheduler | undefined {
+    return this.scheduler;
+  }
+}
